Added RewardValue class and implemented conversion methods

+++ src/main/java/RewardsConverter.java
@@ -14,7 +14,8 @@ public class RewardsConverter {
             return;
         }
         System.out.println("converting $" + input_value + " to miles");
-        var rewardsValue = new RewardValue(cashValue);
+        var rewardsValue = new RewardValue((int) cashValue);
         System.out.println("$" + input_value + " is worth " + rewardsValue.getMilesValue() + " miles");
     }
 }
+
